Add files via upload
Note that You will have to build a new string for the database connection so that it will match the new path the database will have located on your computer/device you may also have to alter the path for the static html website ,so that the shell execute actually works. The code for this is located in a button called website located on the home page. I would also recommend that you have an internet connection as there may be slight inconveniences if you don't(To be more specific the application makes use of a emailing system and I have not included a 'try and except' so that is you why you will get error messages after creating and submitting an application) . There is also the possibility you maybe experience a few  minor bugs but other then that it should work fine.